Refactor account management and update UI components
- Restructured `AccountController` to improve role-based redirects.
- Updated `AdminDashboardController` to use "Administrator" role.
- Removed `User` class from `Users.cs` for simplification.
- Modified `IUserServices` interface and updated method names.
- Enhanced `Login.cshtml` with new layout and styling.
- Changed logout functionality in `Index.cshtml` to use POST.
- Updated database connection string in `appsettings.json`.
