--- Controllers/AccountController .cs
@@ -1,72 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using OmnitakSupportHub.Models;
-using OmnitakSupportHub.Services;
-
-namespace OmnitakSupportHub.Controllers
-{
-    public class AccountController : Controller
-    {
-        private readonly IUserServices _userService;
-
-        public AccountController(IUserServices userService)
-        {
-            _userService = userService;
-        }
-
-        [HttpGet, AllowAnonymous]
-        public IActionResult Login(string? returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            return View();
-        }
-
-        [HttpPost, AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Login(string username, string password, bool rememberMe = false, string? returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-
-            var user = _userService.ValidateUser(username, password);
-            if (user == null)
-            {
-                ModelState.AddModelError(string.Empty, "Invalid credentials");
-                return View();
-            }
-
-            var claims = new List<Claim>
-            {
-                new Claim(ClaimTypes.Name, user.Username),
-                new Claim(ClaimTypes.Role, user.Role)
-            };
-
-            var identity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
-            var principal = new ClaimsPrincipal(identity);
-
-            await HttpContext.SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme, principal,
-                new AuthenticationProperties { IsPersistent = rememberMe });
-
-            if (!string.IsNullOrEmpty(returnUrl) && Url.IsLocalUrl(returnUrl))
-                return Redirect(returnUrl);
-
-            return user.Role == "Admin"
-                ? RedirectToAction("Index", "AdminDashboard")
-                : RedirectToAction("Index", "UserDashboard");
-        }
-
-        [Authorize]
-        public async Task<IActionResult> Logout()
-        {
-            await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
-            return RedirectToAction("Login");
-        }
-
-        [HttpGet]
-        public IActionResult AccessDenied() => View();
-    }
-}
