--- Models/Users.cs
@@ -1,9 +1,0 @@
-﻿namespace OmnitakSupportHub.Models
-{
-    public class User
-    {
-        public string Username { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-        public string Role { get; set; } = string.Empty;
-    }
-}
