--- Services/IUserServices.cs
@@ -1,10 +1,0 @@
-﻿using OmnitakSupportHub.Models;
-
-namespace OmnitakSupportHub.Services
-{
-    public interface IUserServices
-    {
-        User? ValidateUser(string username, string password);
-    }
-}
-
