+++ Views/Account/Login.cshtml
@@ -1,54 +1,56 @@
 ﻿@model OmnitakSupportHub.Models.LoginModel
 @{
-    ViewData["Title"] = "Sign In";
+    ViewData["Title"] = "Omnitak Support Hub";
 }
 
-<div class="row justify-content-center mt-5">
+<div class="row justify-content-center align-items-center min-vh-100 bg-light">
     <div class="col-md-6 col-lg-4">
-        <div class="card shadow-sm border-0">
+        <div class="card border-0 shadow-sm">
             <div class="card-body p-4">
-                <h3 class="card-title text-center mb-4">@ViewData["Title"]</h3>
-                <form asp-action="Login" method="post">
-                    <div asp-validation-summary="All" class="text-danger mb-3"></div>
-                    @Html.AntiForgeryToken()
+                <div class="text-center mb-4">
+                    <div class="rounded-circle bg-success text-white mx-auto mb-2 d-flex align-items-center justify-content-center"
+                         style="width: 48px; height: 48px; font-weight: bold;">
+                        0
+                    </div>
+                    <h4 class="mb-0">Omnitak Support Hub</h4>
+                    <small class="text-muted">IT Support System</small>
+                </div>
 
-                    <div class="form-group mb-3">
-                        <label asp-for="Email" class="form-label">Company Email</label>
+                <form asp-action="Login" method="post">
+                    @Html.AntiForgeryToken()
+                    <div asp-validation-summary="All" class="text-danger small mb-3"></div>
+
+                    <div class="mb-3">
                         <input asp-for="Email"
                                class="form-control"
                                placeholder="Enter your company email" />
-                        <span asp-validation-for="Email"
-                              class="text-danger small"></span>
+                        <span asp-validation-for="Email" class="text-danger small"></span>
                     </div>
 
-                    <div class="form-group mb-2">
-                        <label asp-for="Password" class="form-label">Password</label>
+                    <div class="mb-3">
                         <input asp-for="Password"
-                               type="password"
                                class="form-control"
-                               placeholder="Enter your password" />
-                        <span asp-validation-for="Password"
-                              class="text-danger small"></span>
+                               placeholder="Enter your password"
+                               type="password" />
+                        <span asp-validation-for="Password" class="text-danger small"></span>
                     </div>
 
                     <div class="d-flex justify-content-between align-items-center mb-3">
                         <div class="form-check">
                             <input asp-for="RememberMe" class="form-check-input" />
-                            <label asp-for="RememberMe"
-                                   class="form-check-label small">
+                            <label asp-for="RememberMe" class="form-check-label small">
                                 Remember me
                             </label>
                         </div>
-                        <a asp-action="ForgotPassword"
-                           class="small">Forgot Password?</a>
+                        <a asp-action="ForgotPassword" class="small">Forgot Password?</a>
                     </div>
 
                     <button type="submit"
-                            class="btn btn-primary w-100 mb-3">
+                            class="btn btn-success w-100 fw-semibold">
                         Sign In
                     </button>
 
-                    <div class="text-center small">
+                    <div class="text-center mt-3 small">
                         Don’t have an account?
                         <a asp-action="Register">Create one here</a>
                     </div>
