+++ Views/AdminDashboard/Index.cshtml
@@ -5,4 +5,9 @@
 <h2>Welcome to the Admin Dashboard</h2>
 
 <p>You are logged in as an Admin.</p>
-<a asp-controller="Account" asp-action="Logout">Logout</a>
+
+<form asp-controller="Account" asp-action="Logout" method="post" style="display:inline;">
+    <button type="submit" class="btn btn-link" style="padding:0; border:none; background:none;">
+        Logout
+    </button>
+</form>
